Rework annotations so multiple runners do not clash
An agent should be looking for an annotation that it is advertised for him on
a remote cluster (<local>.wireguard.semaphore.uw.io) and annotate it's node with
info specific for remote watchers (<remote>.wireguard.semaphore.uw.io).

For example a node running on a local cluster "aws" and having configuration for
2 remote clusters named "gcp" and "merit", should create 2 wg keys and 2
endpoints, which will be annotated on the node it is running as:
- gcp.wireguard.semaphore.uw.io/endpoint
- gcp.wireguard.semaphore.uw.io/pubKey
- merit.wireguard.semaphore.uw.io/endpoint
- merit.wireguard.semaphore.uw.io/pubKey

and should be watching remote clusters for the following:
- aws.wireguard.semaphore.uw.io/endpoint
- aws.wireguard.semaphore.uw.io/pubKey

That way we could work with multiple remotes without our runners clashing.