Default watchers resyncPeriod to 0 to disable it.
Using the default 0 duration period should instruct the watchers to avoid
re-listing from cache and getting onUpdate calls.
We should not need this, since our reconcile logic is retrying if an event fails
to update wg peers and never lets go.
https://github.com/kubernetes-sigs/controller-runtime/issues/521#issuecomment-518457302
If a watcher fails to connect to apiservers for a while it should be able to
sync when is back online and create sufficient events to sync our peers again:
https://kubernetes.io/docs/reference/using-api/api-concepts/#efficient-detection-of-changes