Avoid stucking on Runners Run() func
The Run() function waits for node watcher's caches to be synced and can freeze
execution in case of an issue. Avoid holding main on this function since we
allow more than one clusters to be watched in parallel. Also, do not exit if one
of the runners fail to Run(), the healthcheck should be a good enough indicator
to notify us about misbehaviours.L